\name{hw3packageWithRcpp-package}
\alias{hw3packageWithRcpp-package}
\alias{hw3packageWithRcpp}
\docType{package}
\title{
<<<<<<< HEAD
  Multiple Llinear Model and two Metrics
=======
  Multiple Llinear Model and Metric
>>>>>>> a97c631 (two functions)
}
\description{
  This package study multilinear model and related metrics. 
  The function `lrm` give coefficients and fitted value of multilinear models.
<<<<<<< HEAD
  The function mape and R2 compute mean absolute percentage error and R squared value, respectively.
=======
  The function  R2 compute  R squared value.
>>>>>>> a97c631 (two functions)
}
\author{
Yuting Duan

Maintainer: Yuting Duan <yutingd@umich.edu>
}
