% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrm.R
\name{lrm}
\alias{lrm}
\title{Fitting Multinear Regression Model}
\usage{
lrm(X, y)
}
\arguments{
\item{X}{A matrix of all the covariates, the one column no included.}

\item{y}{A vector of observed response.}
}
\value{
A list of coefficients and fitted value.
}
\description{
Give coefficients and fitted value of generalized linear models.
The link function include identity, logit and log function.
}
\examples{
\dontrun{
X = as.matrix(mtcars[,-1])
y = mtcars[,1]
lrm(X, y)
}
}
